+++ common/src/commonMain/kotlin/com/jzbrooks/readlater/common/EntryRepository.kt
@@ -6,5 +6,5 @@ interface EntryRepository {
     val entries: Flow<List<Entry>>
     suspend fun updateEntries()
     suspend fun deleteAllEntries()
-    suspend fun getEntry(id: Long): Entry?
+    suspend fun getEntry(id: Int): Entry?
 }
